add fonts to related language addons
This adds some TTF files in order to have the corresponding character type for upcoming changes in Kodi in the respective language.

The background is that Kodi does not support the characters in very many languages and could only be fixed in a few using "Arial.ttf".
Think about relocating it here, as it is needed in the respective language and so as not to flood Kodi when everyone was in it.